+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [Beta 0.6.1] 2025-02-20
+
+### Fixed
+- 'year' can now be received by /quota.
+
+---
 ## [Beta 0.6.0] 2025-02-19
 
 ### Added